Add device driver support for partial unloading
A driver that supports several profiles (uuid) should not necessarily be
completely unloaded just because one of its profiles has been removed.

This patch adds a new callback that a driver can implement, where
partial removals are handled. The return value of this callback will
tell the core if the driver should be entirely removed (in that case, 0
is returned).