Add partial unloading to device audio driver
This solves the issue fully unloading the driver when one of the
supported profiles is not found in the reverse discovery process.

This can be easily reproduced when both Headset and Gateway are enabled.
When a new phone is paired, the incoming connection can be handled by
the gateway server, which assumes the remote device is a HFP headset,
and installs the HFP_HS_UUID. Once the reverse discovery information is
available, this uuid must be removed, but not the whole driver.