device: Use btd_service to represent profiles
Maintain a list of btd_service pointers instead of btd_profile pointers,
for services that have been probed for this device.

This list will not only reference count the btd_service instances but
also "own" them, in the sense of destroying them (make it completely
unavailable) when the service is removed from the list. This is needed
because the btd_service does not reference count the btd_device
instance.