core: Add btd_server
Add a core object representing a (adapter, profile) pair in a similar
way that btd_service couples a (device, profile) pair. As opposed to
btd_service represents the *remote* support of a profile, btd_server
is introduced to represent a *locally* supported profile.

The new struct should allow removing a lot of boilerplate code in the
profile implementations, such as;
- Maintaining an internal list of probed adapters.
- The agent authorization procedure.
- IO channel handling (listen, shutdown, error-cases, etc.).

Eventually, btd_server is also a step towards supporting multiple local
instances of the same UUID, which could be probed/removed independently.
This extension is however not addressed yet.