service: Add incoming connections with btd_connection
Extend btd_service with helper functions to simplify the implementation
of profiles.

The idea is to introduce btd_connection as an abstraction to couple an
active connection coupling a local profile (btd_server) with a remote
service (btd_service).

The approach follows the pattern currently used for external profiles,
where btd_connection would logically map to struct ext_io (without the
D-Bus part, obviously). The ultimate goal is that the internal profiles
have a similar core infrastructure to simplify their implementation.

btd_service will keep ownership of these connections and therefore the
client code doesn't need to care about details such shutting it down in
case of error etc. If a service is disconnected, all connections will be
automatically closed.

Besides, the agent authorization mechanism is integrated at this point.
btd_service will take care of synchronizing this with the service
discovery as currently done for external profiles.