gateway: Fix bogus disconnected->connected state change
If there's an ongoing connect to gw->sco while we disconnect the through
gateway_close(), we are not really shutting down gw->sco, because it
will be NULL until the callback is called.

If we disconnected we will receive a timeout much later and make the
bogus transition DISCONNECTED->CONNECTED because we try to suspend the
stream instead of closing it.

In the callback we could check the state before suspending the stream,
but there's no reason to wait the timeout and not shutting down the
channel when we disconnected. And this also fix a potential leak of io,
because on connect we were inconditionally incrementing its ref.